Fixed test model
